Update the script to make a CSV file of current MEPs
The old script relied on a large regular expression to pick out data
from the region page, which made it difficult to fix errors when the
page structure changed. This commit introduces a version rewritten in
Python which uses BeautifulSoup to extract data from pages - hopefully
this will be easier to deal with in future.