Fix the start URL and domain for the UK MEP scraper
The start page for this scraper now redirects to a different path and
domain. When the URLs for other pages were being formed, they still used
the old domain, so the pages weren't found. This commit fixes that by
updating the assumed domain that all paths are under, and updates the
path of the start page to it's new location.